Update s3 conf

+++ src/main/kotlin/app/configuration/S3Configuration.kt
@@ -15,7 +15,7 @@ class S3Configuration(
         var clientRegion: String? = "NOT_SET",
         var maxAttempts: Int? = 5,
         var initialBackoffMillis: Long? = 1000L,
-        var backoffMultiplier: Double? = 2.0) {
+        var backoffMultiplier: Long? = 2) {
 
 
     @Bean
